refactor: cleanUp code
