+++ app/src/main/java/com/mashup/ipdam/SingleLiveData.kt
@@ -25,7 +25,7 @@ class SingleLiveEvent<T> : MutableLiveData<T>() {
      */
     @MainThread
     override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
-        super.observe(owner, Observer {
+        super.observe(owner, {
             if (isPending.compareAndSet(true, false)) {
                 observer.onChanged(it)
             }
