+++ app/src/main/res/layout/activity_search.xml
@@ -65,7 +65,8 @@
             app:layout_constraintBottom_toBottomOf="@id/searchView"
             app:layout_constraintEnd_toEndOf="@id/searchView"
             app:layout_constraintTop_toTopOf="@id/searchView"
-            app:srcCompat="@drawable/ic_search" />
+            app:srcCompat="@drawable/ic_search"
+            android:contentDescription="TODO" />
 
 
         <ImageView
