+++ app/src/main/res/layout/bottomsheet_ipdam.xml
@@ -82,7 +82,8 @@
             android:alpha="0"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@drawable/marker" />
+            app:srcCompat="@drawable/marker"
+            android:contentDescription="TODO" />
 
         <TextView
             android:id="@+id/tv_placeNameAndCount"
