+++ app/src/main/res/layout/fragment_profile.xml
@@ -107,7 +107,7 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="18dp"
             android:layout_marginEnd="18dp"
-            android:gravity="right"
+            android:gravity="end"
             android:textColor="@color/surface_color"
             android:textSize="16sp"
             android:textStyle="bold"
