+++ base/src/main/java/com/mashup/base/BaseFragment.kt
@@ -1,11 +1,9 @@
 package com.mashup.base
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
 import androidx.annotation.LayoutRes
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
