+++ base/src/main/java/com/mashup/base/BaseRecyclerView.kt
@@ -1,6 +1,5 @@
 package com.mashup.base
 
-import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.annotation.LayoutRes
